python/requirements.txt
microsoft-kiota-abstractions==1.3.2
microsoft-kiota-http==1.3.1
microsoft-kiota-serialization-json==1.2.0
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-serialization-text==1.0.0
microsoft-kiota-serialization-form==0.1.0
microsoft-kiota-serialization-multipart==0.1.0